[FEAT] postid로 특정 게시글 조회, userid type Long으로 변경

+++ src/main/java/kbsc/kbsc/domain/user/api/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
     */
 
     @GetMapping("/users/{userid}")
-    public Users findUserByIdx(@PathVariable int userid) {
+    public Users findUserByIdx(@PathVariable Long userid) {
         return userDAO.findUserByIdx(userid);
     }
 }
